DUMP, will rewrite later
dumping this for revert's sake.
I will rewrite it so that it starts timer when space is pressed, gets
the time when it is released, and then starts another timer, and then
gets the time when space is pressed again. I hope that this will yield
more accurate rhythm.

+++ include/input.h
@@ -4,7 +4,7 @@
 #include "LSDL.h"
 
 bool handleEvents( SDL_Event e ); //returns whether or not to quit
-void storeInput( float time ); //stores the space presses. takes in the time between each frame
+void storeInput( Uint32 time ); //stores the space presses. takes in the time between each frame
 void startPlayingInput();
 void playInput();
